+++ CHANGELOG.md
@@ -109,3 +109,10 @@ Removed use_in_case2.0 branch from git-workflow filter.
 # 1.2.1
 
 - Added [map](lib/src/map.dart) to library export
+
+
+# 1.3.0
+
+- Modifier [measure](./lib/src/measure.dart) added
+- Modifier [runAtLeast](./lib/src/run_at_least.dart) added
+- Modifier [measuredValue](./lib/src/measure.dart) added